create new classes for MarkCommand and UnmarkCommand which extend EditCommand
