+++ storage/data/panav.txt
@@ -1,2 +1,1 @@
-T ~ 0 ~ workkkkkkk   
-E ~ 0 ~ work    ~ tuesday  ~ wednesday  
+T ~ 1 ~ work
